+++ db/postgresql-async/src/commonMain/kotlin/pw/binom/db/postgresql/async/PreparedStatementMessage.kt
@@ -1,9 +1,0 @@
-package pw.binom.db.postgresql.async
-
-import pw.binom.UUID
-
-abstract class PreparedStatementMessage(
-    val statementId: String,
-    val query: String,
-    val values: List<Any?>,
-):KindedMessage