+++ db/src/commonMain/kotlin/pw/binom/db/SyncPreparedStatement.kt
@@ -1,0 +1,21 @@
+package pw.binom.db
+
+import pw.binom.UUID
+import pw.binom.io.Closeable
+
+interface SyncPreparedStatement : Closeable {
+    val connection: Connection
+    fun set(index: Int, value: Float)
+    fun set(index: Int, value: Int)
+    fun set(index: Int, value: Long)
+    fun set(index: Int, value: String)
+    fun set(index: Int, value: Boolean)
+    fun set(index: Int, value: ByteArray)
+    fun setNull(index: Int)
+    fun executeQuery(): SyncResultSet
+    fun executeUpdate()
+    fun set(index: Int, value: UUID) {
+        val buf = ByteArray(16)
+        set(index, value.toByteArray(buf))
+    }
+}